Move CSV generator to FHIR model
